+++ previous-versions/sync-for-expenses-version-1/src/sdk/sync.ts
@@ -110,22 +110,7 @@ export class Sync {
         });
         const decodedRes = new TextDecoder().decode(httpRes?.data);
         switch (true) {
-            case httpRes?.status == 202:
-                if (utils.matchContentType(contentType, `application/json`)) {
-                    res.syncInitiated = utils.objectToClass(
-                        JSON.parse(decodedRes),
-                        shared.SyncInitiated
-                    );
-                } else {
-                    throw new errors.SDKError(
-                        "unknown content-type received: " + contentType,
-                        httpRes.status,
-                        decodedRes,
-                        httpRes
-                    );
-                }
-                break;
-            case [400, 404, 422].includes(httpRes?.status):
+            case [400, 401, 402, 403, 404, 422, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),