+++ previous-versions/sync-for-expenses-version-1/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [typescript v0.3.0] previous-versions/sync-for-expenses-version-1
 ### Releases
 - [NPM v0.3.0] https://www.npmjs.com/package/@codat/sync-for-expenses-version-1/v/0.3.0 - previous-versions/sync-for-expenses-version-1
+
+## 2023-11-06 01:28:50
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses-v1.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.4.0] previous-versions/sync-for-expenses-version-1
+### Releases
+- [NPM v0.4.0] https://www.npmjs.com/package/@codat/sync-for-expenses-version-1/v/0.4.0 - previous-versions/sync-for-expenses-version-1